// src/api/models/artifact-version-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ArtifactVersionTO
 */
export interface ArtifactVersionTO {
    /**
     * 
     * @type {string}
     * @memberof ArtifactVersionTO
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactVersionTO
     */
    comment?: any;
    /**
     * 
     * @type {number}
     * @memberof ArtifactVersionTO
     */
    milestone?: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactVersionTO
     */
    xml: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactVersionTO
     */
    saveType: ArtifactVersionTOSaveTypeEnum;
    /**
     * 
     * @type {Date}
     * @memberof ArtifactVersionTO
     */
    updatedDate: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactVersionTO
     */
    artifactId: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactVersionTO
     */
    repositoryId: any;
    /**
     * 
     * @type {Array&lt;Deployment&gt;}
     * @memberof ArtifactVersionTO
     */
    deployments?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum ArtifactVersionTOSaveTypeEnum {
    MILESTONE = 'MILESTONE',
    AUTOSAVE = 'AUTOSAVE'
}