// src/api/models/repository-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface RepositoryTO
 */
export interface RepositoryTO {
    /**
     * 
     * @type {string}
     * @memberof RepositoryTO
     */
    id: any;
    /**
     * 
     * @type {string}
     * @memberof RepositoryTO
     */
    name: any;
    /**
     * 
     * @type {string}
     * @memberof RepositoryTO
     */
    description: any;
    /**
     * 
     * @type {number}
     * @memberof RepositoryTO
     */
    existingArtifacts: any;
    /**
     * 
     * @type {number}
     * @memberof RepositoryTO
     */
    assignedUsers: any;
}